requirements.txt
transformers==4.45.1
hydra-core==1.3
hydra_colorlog==1.2.0
torch==2.4.1
datasets==3.0.1
accelerate==0.34.2
flash-attn==2.6.3
bitsandbytes==0.44.1